+++ CHANGELOG.md
@@ -6,9 +6,13 @@
 
 ## 0.3.4
 
-- Added GOMAttribute and GOMNode classes to map GOM attribute and node data.
+- Added GOMAttribute and GOMNode classes to map attributes and nodes.
 
 ## 0.3.5
 
 - Added simple keypath search to GOMNode.
 - Corrected typos on method name.
+
+## 0.3.6
+
+- Using KVC to map dictionaries on GOMNode and GOMAttribute. keypath search uses KVC as well.